% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oopCluster.R
\name{getActiveEvents.mmcluster}
\alias{getActiveEvents.mmcluster}
\title{get active events}
\usage{
\method{getActiveEvents}{mmcluster}(m)
}
\arguments{
\item{m}{model queueing system}
}
\value{
list of active events
}
\description{
get active events
}
