% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oopCluster.R
\name{getNewClocks.mmcluster}
\alias{getNewClocks.mmcluster}
\title{get new clocks}
\usage{
\method{getNewClocks}{mmcluster}(m, e)
}
\arguments{
\item{m}{model queueing system}

\item{e}{list of events}
}
\value{
list of new clocks
}
\description{
get new clocks
}
