% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{getNewClocks}
\alias{getNewClocks}
\title{get new clocks}
\usage{
getNewClocks(m, e)
}
\arguments{
\item{m}{model queueing system}

\item{e}{list of events}
}
\value{
list of new clocks
}
\description{
get new clocks
}
