% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{getRegEst}
\alias{getRegEst}
\title{estimation function}
\usage{
getRegEst(a, alpha)
}
\arguments{
\item{a}{list of accumulated parameters}

\item{alpha}{list of added parameters}
}
\value{
numeric vector of the point estimates
}
\description{
estimation function
}
