% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{getRegEst.statacc}
\alias{getRegEst.statacc}
\title{estimation function}
\usage{
\method{getRegEst}{statacc}(a, alpha = 0.05)
}
\arguments{
\item{a}{list of accumulated parameters}

\item{alpha}{list of added parameters}
}
\value{
numeric vector of the point estimates
}
\description{
estimation function
}
