% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{statacc}
\alias{statacc}
\title{structure of accumulated statistics}
\usage{
statacc(
  l = list(sumY = 0, sumSqY = 0, sumZ = 0, sumSqZ = 0, sumYZ = 0, nReg = 0)
)
}
\arguments{
\item{l}{list of accumulated parameters}
}
\value{
structure of accumulated statistics
}
\description{
structure of accumulated statistics
}
