% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{trace.gsmp}
\alias{trace.gsmp}
\title{building trace for GSMP}
\usage{
\method{trace}{gsmp}(m, numSteps = 1e+05, stopTime = Inf)
}
\arguments{
\item{m}{model queueing system}

\item{numSteps}{amount arrival customers}

\item{stopTime}{modeling time}
}
\value{
numeric list of the accumulated parameters
}
\description{
building trace for GSMP
}
